+++ app/transaction/views.py
@@ -1,15 +1,10 @@
-from .models import Transaction, Budget
+from .models import Transaction
 
 # Create your views here.
-from .serializer import TransactionSerializer, BudgetSerializer
+from .serializer import TransactionSerializer
 from rest_framework import viewsets
 
 
 class TransactionViewSet(viewsets.ModelViewSet):
     queryset = Transaction.objects.all()
     serializer_class = TransactionSerializer
-
-
-class BudgetViewSet(viewsets.ModelViewSet):
-    queryset = Budget.objects.all()
-    serializer_class = BudgetSerializer
